test(pipeline): share BaNCS request fixtures
CodeRabbit (BancsTxnEndpointPicker.test / BancsTxnRequest.test): the
`ACCOUNT_URL`, `origDtBound`, `txnBody`, and `balanceBody` request-shape
builders were duplicated across both suites.

Extract them into a shared `Tests/BancsRequestFixtures.ts` module (the
richer `Ver` / `Operation:'INQ'` variants, accepted by both suites;
`origDtBound` stays private as it is only used via `txnBody`). Both
suites now import the single source, so the fixtures cannot drift as the
BaNCS schema evolves. No behaviour change — 18 tests green.
